refactor: move openapi to a separate crate
