+++ crates/api_models/Cargo.toml
@@ -18,6 +18,7 @@ dummy_connector = ["euclid/dummy_connector", "common_enums/dummy_connector"]
 detailed_errors = []
 payouts = []
 frm = []
+openapi = ["common_enums/openapi"]
 
 [dependencies]
 actix-web = { version = "4.3.1", optional = true }
@@ -36,5 +37,5 @@ cards = { version = "0.1.0", path = "../cards" }
 common_enums = { path = "../common_enums" }
 common_utils = { version = "0.1.0", path = "../common_utils" }
 euclid = { version = "0.1.0", path = "../euclid" }
-masking = { version = "0.1.0", path = "../masking" }
+masking = { version = "0.1.0", path = "../masking", default-features = false, features = ["alloc", "serde"] }
 router_derive = { version = "0.1.0", path = "../router_derive" }
