+++ crates/openapi/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "openapi"
+version = "0.1.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+utoipa = { version = "3.3.0", features = ["preserve_order", "time"] }
+serde_json = "1.0.96"
+
+api_models = { version = "0.1.0", path = "../api_models", features = ["openapi"] }
