+++ crates/openapi/src/lib.rs
@@ -1,0 +1,2 @@
+mod openapi;
+pub mod routes;
