+++ crates/openapi/src/routes.rs
@@ -1,0 +1,7 @@
+#![allow(unused)]
+
+pub mod mandates;
+pub mod merchant_account;
+pub mod merchant_connector_account;
+pub mod payments;
+pub mod refunds;
