+++ crates/openapi/src/routes/merchant_connector_account.rs
@@ -1,0 +1,100 @@
+/// Merchant Connector - Create
+///
+/// Create a new Merchant Connector for the merchant account. The connector could be a payment processor / facilitator / acquirer or specialized services like Fraud / Accounting etc."
+#[utoipa::path(
+    post,
+    path = "/accounts/{account_id}/connectors",
+    request_body = MerchantConnectorCreate,
+    responses(
+        (status = 200, description = "Merchant Connector Created", body = MerchantConnectorResponse),
+        (status = 400, description = "Missing Mandatory fields"),
+    ),
+    tag = "Merchant Connector Account",
+    operation_id = "Create a Merchant Connector",
+    security(("admin_api_key" = []))
+)]
+pub async fn payment_connector_create() {}
+
+/// Merchant Connector - Retrieve
+///
+/// Retrieve Merchant Connector Details
+#[utoipa::path(
+    get,
+    path = "/accounts/{account_id}/connectors/{connector_id}",
+    params(
+        ("account_id" = String, Path, description = "The unique identifier for the merchant account"),
+        ("connector_id" = i32, Path, description = "The unique identifier for the Merchant Connector")
+    ),
+    responses(
+        (status = 200, description = "Merchant Connector retrieved successfully", body = MerchantConnectorResponse),
+        (status = 404, description = "Merchant Connector does not exist in records"),
+        (status = 401, description = "Unauthorized request")
+    ),
+    tag = "Merchant Connector Account",
+    operation_id = "Retrieve a Merchant Connector",
+    security(("admin_api_key" = []))
+)]
+pub async fn payment_connector_retrieve() {}
+
+/// Merchant Connector - List
+///
+/// List Merchant Connector Details for the merchant
+#[utoipa::path(
+    get,
+    path = "/accounts/{account_id}/connectors",
+    params(
+        ("account_id" = String, Path, description = "The unique identifier for the merchant account"),
+    ),
+    responses(
+        (status = 200, description = "Merchant Connector list retrieved successfully", body = Vec<MerchantConnectorResponse>),
+        (status = 404, description = "Merchant Connector does not exist in records"),
+        (status = 401, description = "Unauthorized request")
+    ),
+    tag = "Merchant Connector Account",
+    operation_id = "List all Merchant Connectors",
+    security(("admin_api_key" = []))
+)]
+pub async fn payment_connector_list() {}
+
+/// Merchant Connector - Update
+///
+/// To update an existing Merchant Connector. Helpful in enabling / disabling different payment methods and other settings for the connector etc.
+#[utoipa::path(
+    post,
+    path = "/accounts/{account_id}/connectors/{connector_id}",
+    request_body = MerchantConnectorUpdate,
+    params(
+        ("account_id" = String, Path, description = "The unique identifier for the merchant account"),
+        ("connector_id" = i32, Path, description = "The unique identifier for the Merchant Connector")
+    ),
+    responses(
+        (status = 200, description = "Merchant Connector Updated", body = MerchantConnectorResponse),
+        (status = 404, description = "Merchant Connector does not exist in records"),
+        (status = 401, description = "Unauthorized request")
+    ),
+   tag = "Merchant Connector Account",
+   operation_id = "Update a Merchant Connector",
+   security(("admin_api_key" = []))
+)]
+pub async fn payment_connector_update() {}
+
+/// Merchant Connector - Delete
+///
+/// Delete or Detach a Merchant Connector from Merchant Account
+#[utoipa::path(
+    delete,
+    path = "/accounts/{account_id}/connectors/{connector_id}",
+    params(
+        ("account_id" = String, Path, description = "The unique identifier for the merchant account"),
+        ("connector_id" = i32, Path, description = "The unique identifier for the Merchant Connector")
+    ),
+    responses(
+        (status = 200, description = "Merchant Connector Deleted", body = MerchantConnectorDeleteResponse),
+        (status = 404, description = "Merchant Connector does not exist in records"),
+        (status = 401, description = "Unauthorized request")
+    ),
+    tag = "Merchant Connector Account",
+    operation_id = "Delete a Merchant Connector",
+    security(("admin_api_key" = []))
+)]
+pub async fn payment_connector_delete() {}
