+++ crates/router/Cargo.toml
@@ -95,7 +95,6 @@ tokio = { version = "1.28.2", features = ["macros", "rt-multi-thread"] }
 unicode-segmentation = "1.10.1"
 url = { version = "2.4.0", features = ["serde"] }
 utoipa = { version = "3.3.0", features = ["preserve_order", "time"] }
-utoipa-swagger-ui = { version = "3.1.3", features = ["actix-web"] }
 uuid = { version = "1.3.3", features = ["serde", "v4"] }
 validator = "0.16.0"
 x509-parser = "0.15.0"
@@ -134,6 +133,7 @@ serial_test = "2.0.0"
 time = { version = "0.3.21", features = ["macros"] }
 tokio = "1.28.2"
 wiremock = "0.5"
+
 
 # First party dev-dependencies
 test_utils = { version = "0.1.0", path = "../test_utils" }
