+++ crates/router/src/lib.rs
@@ -12,6 +12,7 @@ pub mod cors;
 pub mod db;
 pub mod env;
 pub(crate) mod macros;
+
 pub mod routes;
 pub mod workflows;
 
@@ -92,15 +93,6 @@ pub fn mk_app(
     >,
 > {
     let mut server_app = get_application_builder(request_body_limit);
-
-    #[cfg(feature = "openapi")]
-    {
-        use utoipa::OpenApi;
-        server_app = server_app.service(
-            utoipa_swagger_ui::SwaggerUi::new("/docs/{_:.*}")
-                .url("/docs/openapi.json", openapi::ApiDoc::openapi()),
-        );
-    }
 
     #[cfg(feature = "dummy_connector")]
     {
